update document identifier from headers on read, propagate change to long for byte length
git-svn-id: svn+ssh://svn.marklogic.com/project/engsvn/client-api/java/trunk@99976 62cac252-8da6-4816-9e9d-6dc37b19578c

+++ src/main/java/com/marklogic/client/io/SearchHandle.java
@@ -188,7 +188,7 @@ public class SearchHandle implements SearchReadHandle<InputStream>, SearchResult
         private Result result = null;
         private MatchLocation[] locations = null;
         private String mimetype = null;
-        private int byteLength = 0;
+        private long byteLength = 0;
 
         public MatchDocumentSummaryImpl(String uri, int score, double confidence, double fitness, String path, Result result) {
             this.uri = uri;
@@ -280,12 +280,12 @@ public class SearchHandle implements SearchReadHandle<InputStream>, SearchResult
         }
 
         @Override
-        public int getByteLength() {
+        public long getByteLength() {
             return byteLength;
         }
 
         @Override
-        public void setByteLength(int length) {
+        public void setByteLength(long length) {
             byteLength = length;
         }
     }
